*DANGEROUS COMMIT* Please read!
Before moving to this commit, you must have already run the `manage.py move_raw_chem_values` command. Once you migrate to 0079, it will not work.
If you have not yet transferred the raw_cas and raw_chem_name values from the Extracted[...]  models to the RawChem model, go back to the previous commit and migration, and run it.